Generisan kod

+++ src/main/java/com/ftn/mbrs/service/impl/KarticaServiceImpl.java
@@ -5,44 +5,55 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.ftn.mbrs.model.Kartica;
+
 import com.ftn.mbrs.model.Vozilo;
-import com.ftn.mbrs.repository.KarticaRepository;
 import com.ftn.mbrs.repository.VoziloRepository;
+
 import com.ftn.mbrs.service.KarticaService;
+import com.ftn.mbrs.model.Kartica;
+import com.ftn.mbrs.repository.KarticaRepository;
 
 @Service
 public class KarticaServiceImpl implements KarticaService{
 
 	@Autowired
 	private KarticaRepository karticaRepository;
-	
+
 	@Autowired
 	private VoziloRepository voziloRepository;
+	
+			
 	
 	@Override
 	public Kartica save(Kartica kartica) {
 		return karticaRepository.save(kartica);
 	}
 
+	//
 	@Override
-	public Kartica update(Kartica kartica) {
-		Kartica temp = karticaRepository.getOne(kartica.getId());
-		temp.setImeVlasnika(kartica.getImeVlasnika());
-		temp.setPrezimeVlasnika(kartica.getPrezimeVlasnika());
-		temp.setKredit(kartica.getKredit());
-		temp.setDate(kartica.getDate());
-		return karticaRepository.save(temp);
+	public Kartica update(Kartica kartica) {		
+		Kartica tempKartica = karticaRepository.getOne(kartica.getId());
+		
+		tempKartica.setImeVlasnika(kartica.getImeVlasnika());    	
+		tempKartica.setPrezimeVlasnika(kartica.getPrezimeVlasnika());    	
+		tempKartica.setKredit(kartica.getKredit());    	
+    	 	
+		
+		return karticaRepository.save(tempKartica);
 	}
 
+	//
 	@Override
-	public void delete(Long id) {
+	public String delete(Long id) {
 		Kartica kartica = karticaRepository.getOne(id);
-		List<Vozilo> vozila = voziloRepository.findByKartica(kartica);
-		for (Vozilo vozilo : vozila) {
-			voziloRepository.delete(vozilo);
-		}
+		
+		List<Vozilo> vozilos = voziloRepository.findByKartica(kartica);
+		if(!vozilos.isEmpty()) {
+			return "ERROR";
+		}    	
+		
 		karticaRepository.deleteById(id);
+		return "OK";
 	}
 
 	@Override
